requirements.txt
Flask==3.0.3
flask-cors==4.0.1
openai==1.35.13
python-dotenv==1.0.1
gTTS==2.5.1